(#3486) - move parseDoc() to own file
This is mostly just code cleanup, but also
partially motivated by performance, because
profiling of map/reduce shows that this function
gets called a lot (once per `emit()` basically).

Below I've noted where I made some micro-optimizations.
